requirements.txt
pandas== 2.0.2
numpy==1.24.3
seaborn==0.12.2
matplotlib==3.7.1
Flask==2.3.2
scikit-learn==1.2.2
xgboost==1.7.5
mysql.connector==2.2.9
cassandra-driver==3.27.0
joblib==1.2.0
gunicorn==19.7.1
-e .